Move manifest related classes to org.robolectric.manifest package.
