+++ robolectric/src/main/java/org/robolectric/res/builder/RobolectricPackageManager.java
@@ -25,10 +25,10 @@ import java.util.Set;
 import java.util.TreeMap;
 import org.robolectric.AndroidManifest;
 import org.robolectric.Robolectric;
-import org.robolectric.res.ActivityData;
-import org.robolectric.res.BroadcastReceiverData;
-import org.robolectric.res.ContentProviderData;
-import org.robolectric.res.IntentFilterData;
+import org.robolectric.manifest.ActivityData;
+import org.robolectric.manifest.BroadcastReceiverData;
+import org.robolectric.manifest.ContentProviderData;
+import org.robolectric.manifest.IntentFilterData;
 import org.robolectric.res.ResName;
 import org.robolectric.res.ResourceIndex;
 import org.robolectric.res.ResourceLoader;
